1.2.00
Added arcing catapult shots. Still a little problematic but that should be fixed at some point.
Reorganized again a bit of code and automated a lot of the systems to the settings.json file for easier ability at modifying units as well as implementing new objects with just its name instead of changing every single sub_key to connect to its settings information. Now it just needs its name ex 'footman' sent to the top class __init__ code and it'll change it for me. Expanded the dev tools a little for easier testing of certain mechanics and units. Other stuff was also probably changed but I can't remember.